backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
billiard==4.2.1
boto3==1.40.27
botocore==1.40.27
build==1.3.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
colorlog==6.9.0
constantly==23.10.4
contourpy==1.3.3
coverage==7.10.6
crcmod==1.7
cryptography==45.0.7
cycler==0.12.1
daphne==4.2.1
dataclasses-json==0.6.7
dill==0.4.0
distro==1.9.0
Django==5.1.5
django-cors-headers==4.8.0
django-db-connection-pool==1.2.6
django-jazzmin==3.0.1
django-redis==6.0.0
django-rename-app==0.1.7
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
doclayout_yolo==0.0.4
durationpy==0.10
et_xmlfile==2.0.0
fast-langdetect==0.2.5
fasttext-predict==0.9.2.4
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
ftfy==6.3.1
gevent==25.8.2
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
json_repair==0.50.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.5.4
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.11
langsmith==0.4.27
lazy_loader==0.4
loguru==0.7.3
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
mineru==2.2.2
mmh3==5.2.0
modelscope==1.29.2
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.22.1
openai==1.107.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
oss2==2.19.1
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pdfminer.six==20250506
pdfplumber==0.11.7
pdftext==0.6.3
pexpect==4.9.0
pillow==11.3.0
pluggy==1.6.0
polars==1.33.1
portalocker==3.2.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==6.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
redis==6.4.0
referencing==0.36.2
regex==2025.9.1
reportlab==4.4.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
robust-downloader==0.0.2
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.1
seaborn==0.13.2
service-identity==24.2.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparams==6.2.0
sqlparse==0.5.3
stringzilla==4.0.0
sympy==1.13.1
tenacity==9.1.2
thop==0.1.1.post2209072238
tifffile==2025.9.9
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.56.1
Twisted==25.5.0
txaio==25.6.1
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.197
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
whitenoise==6.10.0
yarl==1.20.1
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2
zstandard==0.24.0
